"small updated"

+++ app/Http/Controllers/TaskController.php
@@ -26,10 +26,18 @@ class TaskController extends Controller
     }
 
     public function showEdit(Project $project, Task $task){
+        if($project->id != $task->project_id){
+            abort(404, 'Not Found');
+        }
+
         return view('edittask', compact('project', 'task'));
     }
 
     public function edit(Taskrequest $request, Project $project, Task $task){
+        if($project->id != $task->project_id){
+            abort(404, 'Not Found');
+        }
+
         $validated= $request->validated();
         if($validated['status']==='completed'){
             $validated['completed_at']=now();
@@ -41,6 +49,10 @@ class TaskController extends Controller
     }
 
     public function delete(Project $project, Task $task){
+        if($project->id != $task->project_id){
+            abort(404, 'Not Found');
+        }
+        
         $task->delete();
         return redirect()->route('show.project', $project->id)->with('success', 'Task deleted successfully!');
     }
